# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PA School Data API",
    "version": "1.0.0",
    "description": "Read-only JSON API behind chrispeterkins.com/paschools. Pennsylvania PSSA and Keystone results (2013/2015 to present), PVAAS growth, school directory data, and student-weighted aggregates. Responses are cacheable for an hour and carry ETags. Data: Pennsylvania Department of Education, NCES Common Core of Data, U.S. Census Bureau. See /paschools/about for methods and caveats."
  },
  "servers": [
    {
      "url": "https://chrispeterkins.com/paschools/api"
    }
  ],
...
  "components": {
    "parameters": {
      "exam": {
        "name": "exam",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "pssa",
            "keystone"
          ],
          "default": "pssa"
        },
        "description": "PSSA (grades 3-8) or Keystone (grade 11)"
      },
...
      "subject": {
        "name": "subject",
        "in": "query",
        "schema": {
          "type": "string",
          "example": "Mathematics"
        },
        "description": "PSSA: Mathematics, English Language Arts, Science. Keystone: Algebra I, Biology, Literature."
      },
...
    },
...
  },
  "paths": {
...
    "/districts": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Search and page through districts",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "countyId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "District type, e.g. Public, Charter, Cyber Charter, CTC"
          },
          {
            "name": "minEnrollment",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "maxEnrollment",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "countyName",
                "schoolCount",
                "enrollment",
                "proficiency"
              ]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/districts/{id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "One district with its schools and weighted results by year, grade, and subject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/districts/map-values": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "One value per district (keyed by district id and NCES id) for the map's boundary layer",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/exam"
          },
          {
            "$ref": "#/components/parameters/subject"
          },
          {
            "name": "metric",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "proficiency",
                "spending",
                "grad_rate_4yr",
                "low_income"
              ]
            },
            "description": "Boundary value: proficiency (default), spending per pupil, graduation rate, or low-income share (returned in proficiency)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/indicators/district/{id}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "District graduation rate, enrollment-weighted school indicators, enrollment, and AFR spending per pupil by year",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "indicators (with percentiles), enrollment, finance per year vs the state, staff (teachers, students per teacher, average salary and experience vs the state), groups"
          }
        }
      }
    },
...
  }
}